+++ eperusteet-amosaa-service/src/main/java/fi/vm/sade/eperusteet/amosaa/dto/YllapitoDto.java
@@ -3,10 +3,16 @@ package fi.vm.sade.eperusteet.amosaa.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.hibernate.annotations.Cache;
+import org.hibernate.annotations.CacheConcurrencyStrategy;
+
+import javax.persistence.Cacheable;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
+@Cacheable
+@Cache(usage = CacheConcurrencyStrategy.READ_ONLY)
 public class YllapitoDto {
 
     private Long id;
